+++ src/scalar/scalar_pnpn_bc_fctry.f90
@@ -31,12 +31,13 @@
 ! POSSIBILITY OF SUCH DAMAGE.
 !
 !
-!> Defines a factory subroutine for `scalar_pnpn_t`.
+!> Defines boundary condition factory and allocator routines for
+!! `scalar_pnpn_t`.
 submodule(scalar_pnpn) scalar_pnpn_bc_fctry
   use dirichlet, only : dirichlet_t
   use neumann, only : neumann_t
   use user_intf, only : user_t
-  use utils, only : neko_type_error
+  use utils, only : neko_type_error, neko_type_registration_error
   use field_dirichlet, only : field_dirichlet_t
   use field_neumann, only : field_neumann_t
   use expression_dirichlet, only : expression_dirichlet_t
@@ -44,7 +45,8 @@ submodule(scalar_pnpn) scalar_pnpn_bc_fctry
   implicit none
 
   ! List of all possible types created by the boundary condition factories
-  character(len=25) :: SCALAR_PNPN_KNOWN_BCS(6) = [character(len=25) :: &
+  character(len=25), parameter :: SCALAR_PNPN_KNOWN_BCS(6) = [ &
+       character(len=25) :: &
        "dirichlet", &
        "expression_dirichlet", &
        "user_dirichlet", &
@@ -54,14 +56,93 @@ submodule(scalar_pnpn) scalar_pnpn_bc_fctry
 
 contains
 
+  !> Scalar Pn/Pn boundary condition allocator.
+  !! @param[inout] object The object to be allocated.
+  !! @param[in] type_name The name of the boundary condition type.
+  module subroutine scalar_pnpn_bc_allocator(object, type_name)
+    class(bc_t), pointer, intent(inout) :: object
+    character(len=*), intent(in) :: type_name
+    integer :: i
+
+    if (associated(object)) then
+       call object%free()
+       deallocate(object)
+    end if
+
+    select case (trim(type_name))
+    case ("user_dirichlet")
+       allocate(field_dirichlet_t::object)
+    case ("dirichlet")
+       allocate(dirichlet_t::object)
+    case ("expression_dirichlet")
+       allocate(expression_dirichlet_t::object)
+    case ("user_neumann")
+       allocate(field_neumann_t::object)
+    case ("neumann")
+       allocate(neumann_t::object)
+    case ("overset_interface")
+       allocate(overset_interface_t::object)
+    case default
+       do i = 1, scalar_pnpn_bc_registry_size
+          if (trim(type_name) .eq. &
+               trim(scalar_pnpn_bc_registry(i)%type_name)) then
+             call scalar_pnpn_bc_registry(i)%allocator(object)
+             return
+          end if
+       end do
+
+       call neko_type_error("scalar_pnpn boundary conditions", type_name, &
+            SCALAR_PNPN_KNOWN_BCS)
+    end select
+  end subroutine scalar_pnpn_bc_allocator
+
+  !> Register a custom scalar Pn/Pn boundary condition allocator.
+  !! Called in custom user modules inside the `module_name_register_types`
+  !! routine to add a custom type allocator to the registry.
+  !! @param[in] type_name The name of the boundary condition type.
+  !! @param[in] allocator The allocator for the custom user type.
+  module subroutine register_scalar_pnpn_bc(type_name, allocator)
+    character(len=*), intent(in) :: type_name
+    procedure(scalar_pnpn_bc_allocate), pointer, intent(in) :: allocator
+    type(scalar_pnpn_bc_allocator_entry), allocatable :: temp(:)
+    integer :: i
+
+    do i = 1, size(SCALAR_PNPN_KNOWN_BCS)
+       if (trim(type_name) .eq. trim(SCALAR_PNPN_KNOWN_BCS(i))) then
+          call neko_type_registration_error("scalar_pnpn boundary condition", &
+               type_name, .true.)
+       end if
+    end do
+
+    do i = 1, scalar_pnpn_bc_registry_size
+       if (trim(type_name) .eq. &
+            trim(scalar_pnpn_bc_registry(i)%type_name)) then
+          call neko_type_registration_error("scalar_pnpn boundary condition", &
+               type_name, .false.)
+       end if
+    end do
+
+    if (scalar_pnpn_bc_registry_size .eq. 0) then
+       allocate(scalar_pnpn_bc_registry(1))
+    else
+       allocate(temp(scalar_pnpn_bc_registry_size + 1))
+       temp(1:scalar_pnpn_bc_registry_size) = scalar_pnpn_bc_registry
+       call move_alloc(temp, scalar_pnpn_bc_registry)
+    end if
+
+    scalar_pnpn_bc_registry_size = scalar_pnpn_bc_registry_size + 1
+    scalar_pnpn_bc_registry(scalar_pnpn_bc_registry_size)%type_name = type_name
+    scalar_pnpn_bc_registry(scalar_pnpn_bc_registry_size)%allocator => allocator
+  end subroutine register_scalar_pnpn_bc
+
   !> Boundary condition factory. Both constructs and initializes the object.
   !! Will mark a mesh zone for the bc and finalize.
-  !! @param[object] object The boundary condition to be allocated.
+  !! @param[inout] object The boundary condition to be allocated.
   !! @param[in] scheme The `scalar_pnpn` scheme.
   !! @param[inout] json JSON object for initializing the bc.
   !! @param[in] coef SEM coefficients.
   !! @param[in] user The user interface.
-  module subroutine bc_factory(object, scheme, json, coef, user)
+  module subroutine scalar_pnpn_bc_factory(object, scheme, json, coef, user)
     class(bc_t), pointer, intent(inout) :: object
     type(scalar_pnpn_t), intent(in) :: scheme
     type(json_file), intent(inout) :: json
@@ -73,16 +154,11 @@ contains
     character(len=:), allocatable :: default_name
     character(len=64) :: buf
 
-    if (associated(object)) then
-       call object%free()
-       nullify(object)
-    end if
-
     call json_get(json, "type", type)
+    call scalar_pnpn_bc_allocator(object, type)
 
     select case (trim(type))
     case ("user_dirichlet")
-       allocate(field_dirichlet_t::object)
        select type (obj => object)
        type is (field_dirichlet_t)
           obj%update => user%dirichlet_conditions
@@ -90,12 +166,7 @@ contains
           ! solved for.
           call json%add("field_name", scheme%s%name)
        end select
-    case ("dirichlet")
-       allocate(dirichlet_t::object)
-    case ("expression_dirichlet")
-       allocate(expression_dirichlet_t::object)
     case ("user_neumann")
-       allocate(field_neumann_t::object)
        select type (obj => object)
        type is (field_neumann_t)
           obj%update => user%neumann_conditions
@@ -103,17 +174,11 @@ contains
           ! solved for.
           call json%add("field_name", scheme%s%name)
        end select
-    case ("neumann")
-       allocate(neumann_t::object)
     case ("overset_interface")
-       allocate(overset_interface_t::object)
        select type (obj => object)
        type is (overset_interface_t)
           call json%add("field_name", scheme%s%name)
        end select
-    case default
-       call neko_type_error("scalar_pnpn boundary conditions", type, &
-            SCALAR_PNPN_KNOWN_BCS)
     end select
 
     call json_get(json, "zone_indices", zone_indices)
@@ -128,7 +193,11 @@ contains
     object%zone_indices = zone_indices
     call object%finalize()
 
-  end subroutine bc_factory
+    deallocate(type)
+    deallocate(zone_indices)
+    deallocate(default_name)
+
+  end subroutine scalar_pnpn_bc_factory
 
 
 end submodule scalar_pnpn_bc_fctry